adds school boy rugby header component
